🐛 fix(editor/page.tsx): sSR 렌더링 오류 해결
SSR prerender시에 react-quill에서 사용하는 document때문에 build가 안되는 오류를 해결함
